requirements.txt
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
docopt==0.6.2
dulwich==0.22.8
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
installer==0.7.0
ipython==8.12.3
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyring==25.6.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.3
more-itertools==10.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nodeenv==1.9.1
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pbs-installer==2025.7.12
pickleshare==0.7.5
pipreqs==0.5.0
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
pre_commit==4.2.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.1
traitlets==5.14.3
trove-classifiers==2025.5.9.12
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.32.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zstandard==0.23.0
